requirements.txt
requests
sgqlc
pytz
psycopg2-binary
robotframework-requests
robotframework-faker
